Python Version/MusicWallpaperMac/src/requirements.txt
numpy==1.22.4
Pillow==9.1.1
py2app==0.28.2
scikit-learn==1.1.1
scipy==1.8.1
rumps==0.3.0
